add website column to search results and separated university search page into components
